refactor: 그룹 생성 쿼리 적용
