+++ src/pages/MyGroup.jsx
@@ -4,10 +4,26 @@ import CreateGroupModal from '@/components/domain/CreateGroupModal';
 import GroupList from '@/components/domain/GroupList';
 import { COLOR } from '@/styles/color';
 import styled from '@emotion/styled';
-import { useState } from 'react';
+import { useState, useEffect } from 'react';
+import { useGetGroupList } from '@/hooks/queries/group';
+import { useAuthContext } from '@/context/AuthProvider';
 
 function MyGroup() {
   const [createGroupModalVisible, setCreateGroupModalVisible] = useState(false);
+  const groupList = useGetGroupList();
+  const [myGroupList, setMyGroupList] = useState([]);
+  const {
+    authState: { loggedUser },
+  } = useAuthContext();
+
+  useEffect(() => {
+    setMyGroupList(
+      groupList.data?.filter(
+        ({ description }) =>
+          description.master === loggedUser._id || description.member.some((el) => el === loggedUser._id),
+      ),
+    );
+  }, [groupList.data]);
 
   return (
     <StyledPageWrapper>
@@ -19,7 +35,12 @@ function MyGroup() {
             그룹 만들기
           </button>
           {createGroupModalVisible && (
-            <CreateGroupModal visible={createGroupModalVisible} onClose={() => setCreateGroupModalVisible(false)} />
+            <CreateGroupModal
+              visible={createGroupModalVisible}
+              groups={groupList.data}
+              setMyGroupList={setMyGroupList}
+              onClose={() => setCreateGroupModalVisible(false)}
+            />
           )}
         </StyledHeader>
         <StyledDesc>
@@ -33,7 +54,7 @@ function MyGroup() {
           </div>
           <Image src={imgPuzzle} width='30rem' />
         </StyledDesc>
-        <GroupList />
+        <GroupList myGroupList={myGroupList} isLoading={groupList.isLoading} />
       </StyledMyGroup>
     </StyledPageWrapper>
   );
